+++ Scripts/Attributes/AssetPreviewAttribute.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+
+namespace EditorAttributes
+{
+    public class AssetPreviewAttribute : PropertyAttribute
+    {
+	    public float PreviewWidth { get; private set; }
+        public float PreviewHeight { get; private set; }
+
+		/// <summary>
+		/// Attribute to preview an asset in the inspector
+		/// </summary>
+		/// <param name="previewWidth">The width of the preview</param>
+		/// <param name="previewHeight">The height of the preview</param>
+		public AssetPreviewAttribute(float previewWidth = 0f, float previewHeight = 0f)
+        {
+            PreviewWidth = previewWidth;
+            PreviewHeight = previewHeight;
+        }
+    }
+}